refactor(recipe): move recipe to entities
